Get rid of RunMode in favor of hidden clap subcommand
