+++ src/ui/mod.rs
@@ -1,7 +1,55 @@
-pub mod cli;
+mod cli;
 mod fancy_ui;
-pub mod main;
 mod simple_ui;
 mod start;
-pub mod utils;
+mod utils;
 mod writer_tracking;
+
+use std::{fs::File, path::Path, sync::Arc};
+
+pub use self::cli::BurnArgs;
+pub use self::utils::ByteSpeed;
+use crate::{
+    herder::Herder,
+    logging::LogPaths,
+    tty::TermiosRestore,
+    ui::{
+        simple_ui::do_setup_wizard,
+        start::{begin_writing, try_start_burn},
+    },
+};
+use tracing::{debug, info};
+
+pub async fn main(
+    _state_dir: &Path,
+    log_paths: Arc<LogPaths>,
+    args: &BurnArgs,
+) -> anyhow::Result<()> {
+    let _termios_restore = match File::open("/dev/tty") {
+        Ok(tty) => TermiosRestore::new(tty).ok(),
+        Err(error) => {
+            info!(
+                ?error,
+                "failed to open /dev/tty, will not attempt to restore after program"
+            );
+            None
+        }
+    };
+
+    let Some(begin_params) = do_setup_wizard(&args)? else {
+        return Ok(());
+    };
+
+    let mut herder = Herder::new(log_paths.clone());
+    let handle = try_start_burn(
+        &mut herder,
+        &begin_params.make_child_config(),
+        args.root,
+        args.interactive.is_interactive(),
+    )
+    .await?;
+    begin_writing(args.interactive, begin_params, handle, log_paths).await?;
+
+    debug!("Done!");
+    Ok(())
+}
